Really? Reported error indicates the listFiles could have returned null. Does this user have a file where the cache folder should be?
Caused by: java.lang.NullPointerException
at nz.gen.wellington.guardian.android.api.caching.FileService.deleteFiles(FileService.java:94)

+++ src/nz/gen/wellington/guardian/android/api/caching/FileService.java
@@ -91,6 +91,11 @@ public class FileService {
 		}
 		
 		File[] listFiles = cacheDir.listFiles(jsonFilesFilter);
+		if (listFiles == null) {
+			Log.w(TAG, "listFiles was null - indicates cache dir is a file not a folder?");
+			return;
+		}
+		
 		for (int i = 0; i < listFiles.length; i++) {
 			File cacheFile = listFiles[i];
 			if (!cacheFile.delete()) {
